+++ coin.pde
@@ -75,19 +75,17 @@ class Coin extends WorldObject {
         return middle;
     }
     
-    boolean isBallInObject(Ball ball) {
-        /*
-         * Called to determine if the x,y of the provided 'point'
-         * is inside the bounds of this object. Used for collision
-         * detection.
-         */
-         return isCircleInCircle(ball.location, BALL_RADIUS, this.location, this.radius);
-    }
-    
-    void collide() {
+    boolean collide(Ball ball) {
         /*
          * Called when a ball collides with this object
          */
-        ENGINE.world.deleteCoin(this);
+        if (!isCircleInCircle(ball.location, BALL_RADIUS, this.location, this.radius)) {
+            return false;
+        }
+         
+        if (!ball.isSimulated) {
+            ENGINE.world.deleteCoin(this);
+        }
+        return true;
     }
 }
